requirements.txt

Django==4.1.7
djangorestframework==3.14.0
environs==9.5.0

Pillow==9.4.0
python-dotenv==1.0.0
